% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blends-package.R
\docType{package}
\name{blends-package}
\alias{blends}
\alias{blends-package}
\title{blends: Blend Colours and Palettes}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Colour blend functions. These functions make it easier to blend colours and palettes using digital blend modes such as multiply, screen, and overlay.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/davidhodge931/blends}
  \item \url{https://davidhodge931.github.io/blends/}
  \item Report bugs at \url{https://github.com/davidhodge931/blends/issues}
}

}
\author{
\strong{Maintainer}: David Hodge \email{davidhodge931@gmail.com} (\href{https://orcid.org/0000-0002-3868-7501}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item David Hodge \email{davidhodge931@gmail.com} (\href{https://orcid.org/0000-0002-3868-7501}{ORCID}) [copyright holder]
}

}
\keyword{internal}
