% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blends-package.R
\docType{package}
\name{blends-package}
\alias{blends}
\alias{blends-package}
\title{blends: Blend Colour Palettes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Blend colour palettes using blend modes such as multiply and screen.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/davidhodge931/blends}
  \item \url{https://davidhodge931.github.io/blends/}
  \item Report bugs at \url{https://github.com/davidhodge931/blends/issues}
}

}
\author{
\strong{Maintainer}: David Hodge \email{davidhodge931@gmail.com} (\href{https://orcid.org/0000-0002-3868-7501}{ORCID}) [copyright holder]

}
\keyword{internal}
